🚧 feat-wip: Add some flags implementation about cloning and modify run_tasks method.

+++ os/src/syscall/mod.rs
@@ -180,7 +180,7 @@ pub fn syscall(syscall_id: usize, args: [usize; 6]) -> isize {
         SYSCALL_BRK => sys_brk(args[0]),
         SYSCALL_SLEEP => sys_sleep(args[0] as *const TimeVal),
         SYSCALL_GETPPID => sys_getppid(),
-        SYSCALL_WAITPID => sys_waitpid(args[0] as isize, args[1] as *mut i32, args[2]),
+        SYSCALL_WAITPID => sys_waitpid(args[0] as isize, args[1] as *mut i32, args[2] as i32),
         SYSCALL_FORK => sys_fork(args[0], args[1], args[2], args[3], args[4]),
         SYSCALL_YIELD => sys_yield(),
